requirements.txt
arrow==1.2.3
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
configparser==5.3.0
DateTime==4.7
Flask==1.0.2
future==0.18.1
idna==2.8
importlib-metadata==4.12.0
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.14.0
lxml==4.9.1
MarkupSafe==1.1.1
numpy==1.23.3
OpenCC==1.1.4
pandas==1.5.0
pygraphviz==1.5
python-dateutil==2.8.2
python-dotenv==0.10.3
python-time==0.3.0
pytz==2022.2.1
regex==2022.9.13
requests==2.22.0
six==1.12.0
soupsieve==2.3.2.post1
transitions==0.6.9
urllib3==1.25.6
Werkzeug==0.16.0
zipp==3.8.1
zope.interface==5.4.0
